refactor: moved classes into individual files
